+++ client/components/EventsList.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import moment from 'moment'
-import { Event } from '../containers'
+import { Event } from './index'
 
 export class EventsList extends React.Component {
     constructor(props) {
@@ -21,6 +21,7 @@ export class EventsList extends React.Component {
                             <ul className="events-list__list list-view compact-view">
                                 {events.map((event) => (
                                     <Event event={event}
+                                           actions={this.props.actions}
                                            key={event._id}
                                            onClick={this.onEventClick.bind(this, event)} />
                                 ))}
@@ -66,4 +67,5 @@ export class EventsList extends React.Component {
 EventsList.propTypes = {
     onEventClick: React.PropTypes.func,
     events: React.PropTypes.array.isRequired,
+    actions: React.PropTypes.object.isRequired,
 }
