+++ client/components/index.js
@@ -1,5 +1,6 @@
 export { EventsList } from './EventsList'
 export { PlanningItem } from './PlanningItem'
 export { ModalWithForm } from './ModalWithForm'
+export { Event } from './Event'
 import * as fields from './fields'
 export { fields }
