+++ client/components/tests/EventsList_test.jsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import { mount, shallow } from 'enzyme'
 import sinon from 'sinon'
-import { EventsListPanelContainer, Event } from '../../containers'
-import { EventsList } from '../index'
+import { EventsListPanelContainer } from '../../containers'
+import { EventsList, Event } from '../index'
 import { Provider } from 'react-redux'
 import * as actions from '../../actions'
 import { createTestStore } from '../../utils'
@@ -87,7 +87,7 @@ describe('<EventsList />', () => {
     })
     it('trigger an event click', () => {
         const onButtonClick = sinon.spy()
-        const wrapper = shallow(<EventsList events={events} onEventClick={onButtonClick} />)
+        const wrapper = shallow(<EventsList events={events} onEventClick={onButtonClick} actions={{}} />)
         // simulate a click
         wrapper.find(Event).first().simulate('click')
         expect(onButtonClick.calledOnce).toBe(true)
